% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetXFromW}
\alias{GetXFromW}
\title{Function to produce functional dummy variables X from categorical functional data W}
\usage{
GetXFromW(W)
}
\arguments{
\item{W}{2D array, t*n: t is the timestamp and n is the number of the observation}
}
\value{
X 3D array, n*t*Q, Q: the total number of the category
}
\description{
Function to produce functional dummy variables X from categorical functional data W
}
